Fixing the Marshal.GetNativeVariantForObject argument definition for EnumVariantObject
The fix allows users to see the correct accessibility tree of controls that don't support UIA using Inspect tool
Fixes #4120 (LinkLabel accessiblity tree)
Fixes #4121 (CheckedListBox accessiblity tree)
The regression from 0c4d3cb9198d9246

+++ src/System.Windows.Forms/src/System/Windows/Forms/AccessibleObject.EnumVariantObject.cs
@@ -179,7 +179,10 @@ namespace System.Windows.Forms
                 for (i = 0; i < celt && currentChild < childCount; ++i)
                 {
                     ++currentChild;
-                    Marshal.GetNativeVariantForObject(((object)currentChild), GetAddressOfVariantAtIndex(rgVar, i));
+                    // Using "currentChild" as uint type leads to incorrect object boxing and converting an object to a COM VARIANT.
+                    // Because of this, controls without UIA support build an incorrect Accessibility tree.
+                    // It needs to cast "currentChild" to int type before converting to get a correct object argument
+                    Marshal.GetNativeVariantForObject(((object)(int)currentChild), GetAddressOfVariantAtIndex(rgVar, i));
                     Debug.WriteLineIf(CompModSwitches.MSAA.TraceInfo, "AccessibleObject.IEV.Next: adding own child " + currentChild + " of " + childCount);
                 }
